fix: tests location
